+++ src/pl/plajer/villagedefense3/Main.java
@@ -202,7 +202,7 @@ public class Main extends JavaPlugin {
     LanguageManager.init(this);
     saveDefaultConfig();
     if (!(version.equalsIgnoreCase("v1_9_R1") || version.equalsIgnoreCase("v1_10_R1") || version.equalsIgnoreCase("v1_11_R1")
-            || version.equalsIgnoreCase("v1_12_R1") || version.equalsIgnoreCase("v_1_13_R1"))) {
+            || version.equalsIgnoreCase("v1_12_R1") || version.equalsIgnoreCase("v1_13_R1"))) {
       MessageUtils.thisVersionIsNotSupported();
       Bukkit.getConsoleSender().sendMessage(ChatColor.RED + "Your server version is not supported by Village Defense!");
       Bukkit.getConsoleSender().sendMessage(ChatColor.RED + "Sadly, we must shut off. Maybe you consider changing your server version?");