+++ src/pl/plajer/villagedefense3/handlers/InventoryManager.java
@@ -77,7 +77,6 @@ public class InventoryManager {
       invConfig.set("Fire ticks", player.getFireTicks());
       invConfig.set("GameMode", player.getGameMode().toString());
       invConfig.set("Allow flight", player.getAllowFlight());
-      invConfig.set("Scoreboard", player.getScoreboard());
 
       invConfig.set("Size", inventory.getSize());
       invConfig.set("Max stack size", inventory.getMaxStackSize());
@@ -188,7 +187,6 @@ public class InventoryManager {
         player.setGameMode(GameMode.valueOf(invConfig.getString("GameMode")));
         player.setAllowFlight(invConfig.getBoolean("Allow flight"));
         List<String> activePotions = invConfig.getStringList("Active potion effects");
-        player.setScoreboard((Scoreboard) invConfig.get("Scoreboard"));
         for (String potion : activePotions) {
           String[] splited = potion.split("#");
           player.addPotionEffect(new PotionEffect(PotionEffectType.getByName(splited[0]), Integer.valueOf(splited[1]), Integer.valueOf(splited[2])));